+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.0-beta.2] - 2026-05-25
+
 ### Added
 
 - **Per-device equipment buttons**: connect, disconnect, plus device-specific